refactoring logger
