+++ executor/executor.go
@@ -35,13 +35,17 @@ type Executor interface {
 }
 
 func prepareLog(smallSpec *util.CommandSpec, isDebug bool) (err error) {
-	options := []logger.LogOption{}
+	var file io.WriteCloser
+
 	if isDebug {
 		dirPath := filepath.Join(config.SDArtifactsDir, ".sd", "commands")
 		filename := fmt.Sprintf("%v-%v-%v.log", time.Now().Unix(), smallSpec.Namespace, smallSpec.Name)
-		options = append(options, logger.OptDebugWithCreate(dirPath, filename))
+		file, err = createLogFile(dirPath, filename)
+		if err != nil {
+			return err
+		}
 	}
-	lgr, err = logger.New(options...)
+	lgr, err = logger.New(file)
 	return
 }
 
@@ -53,6 +57,23 @@ func parseExecSubCommands(args []string) ([]string, error) {
 		return nil, fmt.Errorf("failed to parse exec args: %w", err)
 	}
 	return f.Args(), nil
+}
+
+func createLogFile(dirPath, filename string) (io.WriteCloser, error) {
+	if filename == "" {
+		filename = fmt.Sprintf("%v.log", time.Now().Unix())
+	}
+	err := os.MkdirAll(dirPath, 0777)
+	if err != nil {
+		return nil, fmt.Errorf("Failed to create logging directory: %v", err)
+	}
+
+	filePath := filepath.Join(dirPath, filename)
+	file, err := os.OpenFile(filePath, os.O_RDWR|os.O_CREATE|os.O_APPEND, 0666)
+	if err != nil {
+		return nil, fmt.Errorf("Failed to create logging file: %v", err)
+	}
+	return file, nil
 }
 
 // New returns each format type of Executor
@@ -136,4 +157,11 @@ func execCommand(path string, args []string) (err error) {
 	state := cmd.ProcessState
 	lgr.Debug.Printf("System Time: %v, User Time: %v\n", state.SystemTime(), state.UserTime())
 	return
+}
+
+// CleanUp close file you use.
+func CleanUp() {
+	if lgr.File() != nil {
+		lgr.File().Close()
+	}
 }
