+++ go.mod
@@ -7,5 +7,6 @@ require (
 	github.com/pkg/errors v0.9.1
 	github.com/stretchr/testify v1.2.2
 	golang.org/x/crypto v0.0.0-20200214034016-1d94cc7ab1c6
+	golang.org/x/sys v0.0.0-20201009025420-dfb3f7c4e634 // indirect
 	gopkg.in/yaml.v2 v2.3.0
 )
